+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 ---
 
  * Remove `Bic::INVALID_BANK_CODE_ERROR` constant. This error code was not used in the Bic constraint validator anymore.
+ * Remove `Expression::getDefaultOption()`, use named arguments instead.
+ * Remove `Expression::etRequiredOptions()`, use named arguments instead.
 
 7.4
 ---